requirements.txt
wheel
aiohttp
aioredis
redis
hiredis
b2sdk
async-timeout
pillow
gql[all]
sqlalchemy[asyncio]
alembic
psycopg2
pottery
httpx
celery[redis]
gevent
sentry-sdk